AutoCHANNEL partial implementation
