+++ src/daf/__init__.py
@@ -10,7 +10,6 @@ from .dtypes import *
 from .guild import *
 from .message import *
 from .logging import *
-from .gen import *
 
 from .misc import DOCUMENTATION_MODE
 if DOCUMENTATION_MODE:
